requirements.txt
psycopg2-binary>=2.9.9
python-dotenv>=1.0.0
faker>=20.0.0
streamlit>=1.28.0
pandas>=2.0.0
plotly>=5.17.0
